"ads2_m6_a1_1: 0/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/Module 6/m6/ADS-2Week-1/Solution.java
@@ -14,7 +14,7 @@ class PageRank {
 			indegrees[i] = dg.indegree(i);
 			for (int j = 0; j < indegrees[i]; j++) {
 				for (int k = 0; k < 1000; k++) {
-					pagerank[i] /= initalpr / dg.outdegree(2);
+					pagerank[i] /= (double)initalpr / dg.outdegree(2);
 				}
 			}
 			System.out.println(pagerank[i]);
